IO - add new writer worker
the osc out prescheduler just deals with timers, and then sends all OSC for scsynth to the new osc writer which maintains a single queue and is the only thread that may write to the ring buffer. This simplifies the writing to make it easier to observe and debug in addition to removing any potential concurrency issues.